Implement purge command
This command looks through the current working directory and deletes
any file that is duplicated in another directory passed to the command.
Hashes are used to determine if two files are equal. As a precaution,
if a hash in a Hashes.xml file is stale, the file is ignored to make
sure a file that was modified and not stored anywhere else isn't
deleted just because the Hashes.xml wasn't up to date.